+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Here is a template for new release sections
 - New module `exceptions.py` in `multi_vector_simulator.utils` to gather custom MVS exceptions (#656)
 - New argument for functions `E1.convert_demand_to_dataframe`, `F1.plot_timeseries`, `F2.ready_timeseries_plots` (#665)
 - File .github/workflow/main.yml for github actions (#668)
+
 - `energyBusses` now have to be defined by the user via `energyBusses.csv` (#649)
 - Input validation test `C1.check_for_label_duplicates` (#649)
 - Constant variables: `JSON_PROCESSED`, `JSON_WITH_RESULTS`, `JSON_FILE_EXTENSION` (#649)
@@ -38,7 +39,7 @@ Here is a template for new release sections
 - "SOC" as string representative in `utils/constants.py`, used in `F1` and `E0` (#666)
 - SOC plot of storages is added to the autoreport (#666)
 - Test for correct storage labelling in `A1.add_storage_components()` (#666)
- - Test for getting two time series with `E1.get_timeseries_per_bus()` for storage (input and output power) also if storage is directly connected to bus (#666)
+- Test for getting two time series with `E1.get_timeseries_per_bus()` for storage (input and output power) if storage is directly connected to bus (#666)
 
 ### Changed
 - Function `utils.compare_input_parameters_with_reference` accepts parameters as dict for json comparison (#656)